Create a Serializer to serialize values
The Serializer keeps track of alignment (using a CountingWriter under
the hood), so that the caller doesn't have to keep track of it. Also
the target endianness is part of the type, so that we can test both
little- and big-endian serializers, but still only have code for native
serialization in actual applications.